+++ packages/headless/src/controllers/product-listing/range-facet/numeric-facet/headless-product-listing-numeric-facet.ts
@@ -3,9 +3,9 @@ import {ProductListingEngine} from '../../../../app/product-listing-engine/produ
 import {
   logFacetClearAll,
   logFacetUpdateSort,
-} from '../../../../features/facets/facet-set/facet-set-analytics-actions';
+} from '../../../../features/facets/facet-set/facet-set-product-listing-analytics-actions';
 import {RangeFacetSortCriterion} from '../../../../features/facets/range-facets/generic/interfaces/request';
-import {getAnalyticsActionForToggleRangeFacetSelect} from '../../../../features/facets/range-facets/generic/range-facet-utils';
+import {getLegacyAnalyticsActionForToggleRangeFacetSelect} from '../../../../features/facets/range-facets/generic/range-facet-utils';
 import {NumericRangeRequest} from '../../../../features/facets/range-facets/numeric-facet-set/interfaces/request';
 import {NumericFacetValue} from '../../../../features/facets/range-facets/numeric-facet-set/interfaces/response';
 import {numericFacetSetReducer as numericFacetSet} from '../../../../features/facets/range-facets/numeric-facet-set/numeric-facet-set-slice';
@@ -78,7 +78,10 @@ export function buildNumericFacet(
       coreController.toggleSelect(selection);
       dispatch(fetchProductListing()).then(() =>
         dispatch(
-          getAnalyticsActionForToggleRangeFacetSelect(getFacetId(), selection)
+          getLegacyAnalyticsActionForToggleRangeFacetSelect(
+            getFacetId(),
+            selection
+          )
         )
       );
     },