+++ packages/headless/src/features/facets/facet-set/facet-set-analytics-actions.ts
@@ -49,6 +49,7 @@ export interface LogFacetUpdateSortActionCreatorPayload {
   criterion: FacetSortCriterion | RangeFacetSortCriterion;
 }
 
+//TODO: KIT-2859
 export const logFacetUpdateSort = (
   payload: LogFacetUpdateSortActionCreatorPayload
 ): SearchAction =>
@@ -69,6 +70,7 @@ export const logFacetUpdateSort = (
     return client.makeFacetUpdateSort(metadata);
   });
 
+//TODO: KIT-2859
 export const logFacetClearAll = (facetId: string): SearchAction =>
   makeAnalyticsAction('analytics/facet/reset', (client, state) => {
     validatePayload(facetId, facetIdDefinition);
@@ -91,6 +93,7 @@ export interface LogFacetSelectActionCreatorPayload {
   facetValue: string;
 }
 
+//TODO: KIT-2859
 export const logFacetSelect = (
   payload: LogFacetSelectActionCreatorPayload
 ): SearchAction =>
@@ -109,6 +112,7 @@ export const logFacetSelect = (
     return client.makeFacetSelect(metadata);
   });
 
+//TODO: KIT-2859
 export const logFacetExclude = (
   payload: LogFacetSelectActionCreatorPayload
 ): SearchAction =>
@@ -139,6 +143,7 @@ export interface LogFacetDeselectActionCreatorPayload {
   facetValue: string;
 }
 
+//TODO: KIT-2859
 export const logFacetDeselect = (
   payload: LogFacetDeselectActionCreatorPayload
 ): SearchAction =>
@@ -168,6 +173,7 @@ export interface LogFacetBreadcrumbActionCreatorPayload {
   facetValue: string;
 }
 
+//TODO: KIT-2859
 export const logFacetBreadcrumb = (
   payload: LogFacetBreadcrumbActionCreatorPayload
 ): SearchAction =>